+++ src/frontend/packages/lib_video/src/components/common/VideoWidgetProvider/index.spec.tsx
@@ -225,6 +225,9 @@ describe('<VideoWidgetProvider />', () => {
     //  Closed captions
     expect(screen.getByText('Closed captions')).toBeInTheDocument();
 
+    // License Manager
+    expect(screen.getByText('License')).toBeInTheDocument();
+
     // visibility
     expect(
       screen.getByText('Visibility and interaction parameters'),